chore: Use Node v24
Unfortunately, we previously thought the issue with CI failing had to do with
our `package-lock.json` file, so we blew it up with `node_modules` and
re-installed. However, the issue was actually because we were using different
versions of Node.js between local development and CI. :\ Issue is fixed now.