chore: 1st Draft of "Root" Combobox Documentation Pages
This is as far as we'll go for these. Now we can start tackling
the more specific documentation pages. If we feel it's truly
necessary, we can wrap back around and provide additional updates
to these "root pages". We still need to update `src/README.md`
at some point as well.