refactor: Add an explicit `Properties` section to `ComboboxListbox` docs
It's technically useless, but it will allow us to establish a consistent
pattern across all of our Custom Elements. This will enables devs to KNOW
that this component has no special properties of its own (instead of devs
having to infer/guess that from the lack of a `Properties` section).