fix: Don't register Custom Elements in their corresponding files (docs)
The order in which the Custom Elements for the `Combobox` component are
registered matters. Thus, calling `customElements.define()` within a
Custom Element's class definition file in our documentation might be
a footgun for some people. We have chosen to remove this footgun. Other
parts of our documentation make it clear how to register extensions of
our Custom Elements (and in what order to do so); that should be sufficient.