feat: Enable `string`-coercion of `ComboboxField`/`ComboboxOption` values
This ultimately leads to a better DX. No developer wants to coerce every
numeric value to a pure string.

Note that the `ComboboxOption` already supported this functionality, but
a new test has been added for it to guarantee that this functionality is
never lost. Also, the types have been updated to support this behavior.
The _raw_ `ComboboxField` and `ComboboxOption` value properties should
still only accept `string`s _in TS_, just like `<select>`/`<option>`.