fix: Properly handle `valueis` transitions when no `options` are present
This has been a long-known bug within the `Combobox` component, but it generally
hasn't been practical to address. In this case, it still might not be practical
to address. However, I don't want to encounter obscure bug reports as a result
of this bug. So, since I have some extra time, I figured I'd just knock out the
bug. Now, `valueis` should be _guaranteed_ to transition in a way that keeps the
`combobox` in a valid state at all times. (If not, we at least have better testing
in place to improve our transitions in the future now. But we feel pretty
confident in the bug fix's implementation.)