chore: Make Form CSS more usable in `Safari`
Colors like `fieldtext` and `field` worked in `Chrome` and `Firefox`,
but not `Safari`. It's only for quick prototyping, so it's not super
important; but people probably at least want to be able to _see_ what
our library is doing without having to modify any CSS files in our
StackBlitz demos.