fix: Only perform `Combobox` ID synchronization if component is mounted
The `Combobox` ID synchronization logic requires the owned `listbox` and `option`
elements to already be accessible to the `ComboboxField`. However, this may not
be the case if the entire component group hasn't been mounted to the DOM yet. In
such a case, we should do nothing so that errors won't be unexpectedly thrown.

Remember that the `SelectEnhancer` takes responsibility for synchronizing the IDs
on mount / initial page load. So this is a completely acceptable change.