chore: Add some examples of async option loading for the `Combobox`
We need to put these examples on StackBlitz. But being able to have
and test them locally will do for now. It seems that Svelte and Vue
struggle a bit in this area, presumably due to how they handle
reconciliation when rendering components/elements to the DOM. We
should probably raise GitHub Issues when the opportunity arises.

Because frameworks render DOM nodes differently, `anyvalue` will
typically be easier to use for async loading than the other modes
(as we anticipated). We've clarified this in the documentation as well.
In the examples, we should probably clarify that all of the logic can
be abstracted away into helper functions so that they really only have
to worry creating async loader functions and nothing else.

Additionally, it seems that Solid prefers to spread properties rather
than attributes, so we added some extra types to make it easifer for
people to use the `valueis` attribute on the `<Select>` component.